requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.28
langchain-core==0.3.15
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.0.8
supabase==2.7.4
